divided templates standalone designer to shell to display local placeholder
